Rename `commands.rs` -> `command.rs`
