+++ crates/sncast/src/helpers/command.rs
