+++ crates/sncast/src/helpers/mod.rs
@@ -1,7 +1,7 @@
 pub mod account;
 pub mod block_explorer;
 pub mod braavos;
-pub mod commands;
+pub mod command;
 pub mod config;
 pub mod configuration;
 pub mod constants;
